+++ isitfit/cli/tags.py
@@ -3,7 +3,7 @@ logger = logging.getLogger('isitfit')
 
 import click
 
-from ..utils import display_footer
+from ..utils import IsitfitCommand
 
 @click.group(help="Explore EC2 tags", invoke_without_command=False)
 def tags():
@@ -11,9 +11,10 @@ def tags():
 
 
 
-@tags.command(help="Generate new tags suggested by isitfit for each EC2 instance")
+@tags.command(help="Generate new tags suggested by isitfit for each EC2 instance", cls=IsitfitCommand)
 @click.option('--advanced', is_flag=True, help='Get advanced suggestions of tags. Requires login')
-def suggest(advanced):
+@click.pass_context
+def suggest(ctx, advanced):
   # gather anonymous usage statistics
   from ..utils import ping_matomo, IsitfitCliError
   ping_matomo("/tags/suggest")
@@ -21,67 +22,49 @@ def suggest(advanced):
   tl = None
   if not advanced:
     from ..tags.tagsSuggestBasic import TagsSuggestBasic
-    tl = TagsSuggestBasic()
+    tl = TagsSuggestBasic(ctx)
   else:
     from ..tags.tagsSuggestAdvanced import TagsSuggestAdvanced
-    tl = TagsSuggestAdvanced()
+    tl = TagsSuggestAdvanced(ctx)
 
-  try:
-    tl.prepare()
-    tl.fetch()
-    tl.suggest()
-    tl.display()
-  except IsitfitCliError as e:
-    logger.error("Error: %s"%str(e))
-    import sys
-    sys.exit(1)
-
-  display_footer()
+  tl.prepare()
+  tl.fetch()
+  tl.suggest()
+  tl.display()
 
 
-@tags.command(help="Dump existing EC2 tags in tabular form into a csv file")
-def dump():
+@tags.command(help="Dump existing EC2 tags in tabular form into a csv file", cls=IsitfitCommand)
+@click.pass_context
+def dump(ctx):
   # gather anonymous usage statistics
   from ..utils import ping_matomo, IsitfitCliError
   ping_matomo("/tags/dump")
 
   from ..tags.tagsDump import TagsDump
-  tl = TagsDump()
+  tl = TagsDump(ctx)
 
-  try:
-    tl.fetch()
-    tl.suggest() # not really suggesting. Just dumping to csv
-    tl.display()
-  except IsitfitCliError as e:
-    logger.error("Error: %s"%str(e))
-    import sys
-    sys.exit(1)
-
-  display_footer()
+  tl.fetch()
+  tl.suggest() # not really suggesting. Just dumping to csv
+  tl.display()
 
 
 
-@tags.command(help="Push EC2 tags from csv file")
+@tags.command(help="Push EC2 tags from csv file", cls=IsitfitCommand)
 @click.argument('csv_filename') #, help='Path to CSV file holding tags to be pushed. Should match format from `isitfit tags dump`')
 @click.option('--not-dry-run', is_flag=True, help='True for dry run (simulated push)')
-def push(csv_filename, not_dry_run):
+@click.pass_context
+def push(ctx, csv_filename, not_dry_run):
   # gather anonymous usage statistics
   from ..utils import ping_matomo, IsitfitCliError
   ping_matomo("/tags/push")
 
   from ..tags.tagsPush import TagsPush
 
-  tp = TagsPush(csv_filename)
-  try:
-    tp.read_csv()
-    tp.validateTagsFile()
-    tp.pullLatest()
-    tp.diffLatest()
-    tp.processPush(not not_dry_run)
-  except IsitfitCliError as e:
-    logger.error("Error: %s"%str(e))
-    import sys
-    sys.exit(1)
+  tp = TagsPush(csv_filename, ctx)
 
-  display_footer()
+  tp.read_csv()
+  tp.validateTagsFile()
+  tp.pullLatest()
+  tp.diffLatest()
+  tp.processPush(not not_dry_run)